+++ Sources/Sentry/SentryHub.m
@@ -256,7 +256,7 @@ SentryHub ()
                        withScope:(SentryScope *)scope
          additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems
 {
-    SentrySampleDecision decision = transaction.trace.context.sampled;
+    SentrySampleDecision decision = transaction.trace.sampled;
     if (decision != kSentrySampleDecisionYes) {
         [self.client recordLostEvent:kSentryDataCategoryTransaction
                               reason:kSentryDiscardReasonSampleRate];
@@ -363,6 +363,20 @@ SentryHub ()
                        customSamplingContext:customSamplingContext];
 }
 
+- (SentryTransactionContext *)transactionContext:(SentryTransactionContext *)context
+                                     withSampled:(SentrySampleDecision)sampleDecision
+{
+
+    return [[SentryTransactionContext alloc] initWithName:context.name
+                                               nameSource:context.nameSource
+                                                operation:context.operation
+                                                  traceId:context.traceId
+                                                   spanId:context.spanId
+                                             parentSpanId:context.parentSpanId
+                                                  sampled:sampleDecision
+                                            parentSampled:context.parentSampled];
+}
+
 - (id<SentrySpan>)startTransactionWithContext:(SentryTransactionContext *)transactionContext
                                   bindToScope:(BOOL)bindToScope
                               waitForChildren:(BOOL)waitForChildren
@@ -373,7 +387,8 @@ SentryHub ()
                                             customSamplingContext:customSamplingContext];
 
     SentryTracesSamplerDecision *samplerDecision = [_tracesSampler sample:samplingContext];
-    transactionContext.sampled = samplerDecision.decision;
+    transactionContext = [self transactionContext:transactionContext
+                                      withSampled:samplerDecision.decision];
     transactionContext.sampleRate = samplerDecision.sampleRate;
 
     SentryProfilesSamplerDecision *profilesSamplerDecision =
@@ -401,7 +416,8 @@ SentryHub ()
                                             customSamplingContext:customSamplingContext];
 
     SentryTracesSamplerDecision *samplerDecision = [_tracesSampler sample:samplingContext];
-    transactionContext.sampled = samplerDecision.decision;
+    transactionContext = [self transactionContext:transactionContext
+                                      withSampled:samplerDecision.decision];
     transactionContext.sampleRate = samplerDecision.sampleRate;
 
     SentryProfilesSamplerDecision *profilesSamplerDecision =