+++ Sources/Sentry/SentrySpanContext.m
@@ -4,13 +4,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface
-SentrySpanContext () {
-    NSMutableDictionary<NSString *, NSString *> *_tags;
-}
-
-@end
-
 @implementation SentrySpanContext
 
 - (instancetype)initWithOperation:(NSString *)operation
@@ -33,14 +26,28 @@ SentrySpanContext () {
                       operation:(NSString *)operation
                         sampled:(SentrySampleDecision)sampled
 {
+    return [self initWithTraceId:traceId
+                          spanId:spanId
+                        parentId:parentId
+                       operation:operation
+                 spanDescription:nil
+                         sampled:sampled];
+}
+
+- (instancetype)initWithTraceId:(SentryId *)traceId
+                         spanId:(SentrySpanId *)spanId
+                       parentId:(nullable SentrySpanId *)parentId
+                      operation:(NSString *)operation
+                spanDescription:(nullable NSString *)description
+                        sampled:(SentrySampleDecision)sampled
+{
     if (self = [super init]) {
         _traceId = traceId;
         _spanId = spanId;
         _parentSpanId = parentId;
-        self.sampled = sampled;
-        self.operation = operation;
-        self.status = kSentrySpanStatusUndefined;
-        _tags = [[NSMutableDictionary alloc] init];
+        _sampled = sampled;
+        _operation = operation;
+        _spanDescription = description;
     }
     return self;
 }
@@ -53,26 +60,6 @@ SentrySpanContext () {
     return type;
 }
 
-- (NSDictionary<NSString *, NSString *> *)tags
-{
-    @synchronized(_tags) {
-        return _tags.copy;
-    }
-}
-- (void)setTagValue:(NSString *)value forKey:(NSString *)key
-{
-    @synchronized(_tags) {
-        [_tags setValue:value forKey:key];
-    }
-}
-
-- (void)removeTagForKey:(NSString *)key
-{
-    @synchronized(_tags) {
-        [_tags removeObjectForKey:key];
-    }
-}
-
 - (NSDictionary<NSString *, id> *)serialize
 {
     NSMutableDictionary *mutabledictionary = @{
@@ -82,12 +69,6 @@ SentrySpanContext () {
         @"op" : self.operation
     }
                                                  .mutableCopy;
-
-    @synchronized(_tags) {
-        if (_tags.count > 0) {
-            mutabledictionary[@"tags"] = _tags.copy;
-        }
-    }
 
     // Since we guard for 'undecided', we'll
     // either send it if it's 'true' or 'false'.
@@ -101,10 +82,6 @@ SentrySpanContext () {
 
     if (self.parentSpanId != nil) {
         [mutabledictionary setValue:self.parentSpanId.sentrySpanIdString forKey:@"parent_span_id"];
-    }
-
-    if (self.status != kSentrySpanStatusUndefined) {
-        [mutabledictionary setValue:nameForSentrySpanStatus(self.status) forKey:@"status"];
     }
 
     return mutabledictionary;