+++ Tests/SentryTests/SentryHubTests.swift
@@ -216,7 +216,7 @@ class SentryHubTests: XCTestCase {
         let span = fixture.getSut().startTransaction(name: fixture.transactionName, operation: fixture.transactionOperation)
         let tracer = Dynamic(span)
         XCTAssertEqual(tracer.transactionContext.name, fixture.transactionName)
-        XCTAssertEqual(span.context.operation, fixture.transactionOperation)
+        XCTAssertEqual(span.operation, fixture.transactionOperation)
     }
     
     func testStartTransactionWithContext() {
@@ -227,7 +227,7 @@ class SentryHubTests: XCTestCase {
         
         let tracer = Dynamic(span)
         XCTAssertEqual(tracer.transactionContext.name, fixture.transactionName)
-        XCTAssertEqual(span.context.operation, fixture.transactionOperation)
+        XCTAssertEqual(span.operation, fixture.transactionOperation)
     }
 
     func testStartTransactionWithNameSource() {
@@ -240,7 +240,7 @@ class SentryHubTests: XCTestCase {
         let tracer = Dynamic(span)
         XCTAssertEqual(tracer.transactionContext.name, fixture.transactionName)
         XCTAssertEqual(tracer.transactionContext.nameSource, SentryTransactionNameSource.url)
-        XCTAssertEqual(span.context.operation, fixture.transactionOperation)
+        XCTAssertEqual(span.operation, fixture.transactionOperation)
     }
     
     func testStartTransactionWithContextSamplingContext() {
@@ -257,7 +257,7 @@ class SentryHubTests: XCTestCase {
         let tracer = Dynamic(span)
         XCTAssertEqual(tracer.transactionContext.name, fixture.transactionName)
         XCTAssertEqual(customSamplingContext?["customKey"] as? String, "customValue")
-        XCTAssertEqual(span.context.operation, fixture.transactionOperation)
+        XCTAssertEqual(span.operation, fixture.transactionOperation)
     }
     
     func testStartTransaction_checkContextSampleRate_fromOptions() {
@@ -265,7 +265,7 @@ class SentryHubTests: XCTestCase {
         options.tracesSampleRate = 0.49
         
         let span = fixture.getSut().startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), customSamplingContext: ["customKey": "customValue"])
-        let context = span.context as? TransactionContext
+        let context = (span as? SentryTracer)?.transactionContext
         
         XCTAssertEqual(context?.sampleRate, 0.49)
     }
@@ -277,7 +277,7 @@ class SentryHubTests: XCTestCase {
         }
         
         let span = fixture.getSut().startTransaction(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), customSamplingContext: ["customKey": "customValue"])
-        let context = span.context as? TransactionContext
+        let context = (span as? SentryTracer)?.transactionContext
         
         XCTAssertEqual(context?.sampleRate, 0.51)
     }
@@ -322,7 +322,7 @@ class SentryHubTests: XCTestCase {
         Dynamic(hub).sampler.random = fixture.random
         
         let span = hub.startTransaction(name: fixture.transactionName, operation: fixture.transactionOperation)
-        XCTAssertEqual(span.context.sampled, .no)
+        XCTAssertEqual(span.sampled, .no)
     }
 
     func testCaptureSampledTransaction_ReturnsEmptyId() {
@@ -860,6 +860,6 @@ class SentryHubTests: XCTestCase {
 
         let span = hub.startTransaction(name: fixture.transactionName, operation: fixture.transactionOperation)
 
-        XCTAssertEqual(expected, span.context.sampled)
+        XCTAssertEqual(expected, span.sampled)
     }
 }