+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 This version introduces a dependency on Swift, which only impacts you if you don't use Swift. If your project uses Swift already, this change will not affect you.
 
+### Breaking Changes
+
+- Make SpanContext immutable (#)
+    - Remove tags from SpanContext 
+    - Remove context property from SentrySpan
+
 ### Features
 
 - Properly demangle Swift class name (#2162)